refactor(auth): compute OAuth provider presence server-side (drop public endpoint)
Replace the client-fetched /api/auth/social-providers route with a server
component that reads provider presence (booleans only, never secrets) and
passes it to the login form as a prop. Removes the unauthenticated public
endpoint, eliminates the client round-trip + loading flash, and renders the
correct enabled/disabled buttons on first paint.

- apps/web/app/(auth)/login/page.tsx: now a server component; reads env presence
- apps/web/app/(auth)/login/login-content.tsx: client form, takes socialEnabled prop
- apps/web/app/api/auth/social-providers/route.ts: removed
